requirements.txt
# Core dependencies
discord.py>=2.3.0
wavelink>=3.4.0

# Database
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0

# Utilities
psutil>=5.9.0

# Optional: For lyrics support (uncomment if needed)
# lyricsgenius>=3.0.1

# Optional: For better async HTTP (recommended)
aiohttp>=3.9.0
